requirements.txt
bar-chart-race==0.1.0
pandas==1.3.3
pyaml==21.10.1
openpyxl==3.0.9